Move things around a little. Metrics stuff in the metrics module. Tracing stuff in the tracing module.
