+++ apollo-router/src/plugins/telemetry/metrics/prometheus.rs
@@ -1,6 +1,6 @@
 use crate::future::BoxFuture;
 use crate::plugins::telemetry::config::MetricsCommon;
-use crate::plugins::telemetry::{MetricsBuilder, MetricsConfigurator};
+use crate::plugins::telemetry::metrics::{MetricsBuilder, MetricsConfigurator};
 use apollo_router_core::{http_compat, ResponseBody};
 use bytes::Bytes;
 use http::StatusCode;
